New Dropbox Importer + Oracle Connections!
Today I made a new script that does the following:

-Connect to the Oracle database and get a list of all the .xml files
that need to be imported to the system.
-Take the BLOB version of that file and make a physical file of it.
-Import the physical file with the cerberus API thing that we worked on
a few weeks ago with the API login.

The script works pretty well so far - the only hitch is that the curl
connection that actually imports the .xml files does not work.

On a more fundamental change, we can now actually connect to Oracle.
This makes use of php/dao/_util.php, sec/oracle.php (as the proof of
concept) and sec/php/data/rec/sql/dao/Dao.php