Correct Warren Changes + Dao Oracle Compatibility + CCD Procedure
-process.php now works as the CCD inloader and logs its progress.
Theoretically the whole procedure works.

-Dao-Oracle-support - this fully supports Oracle as well as MySQL now.
The reason why it's not live is because there are a few other changes
that are being worked on in Dao.php right now. But this file is a safe
rollback.

-Procedures and SqlRec reflect the SQL identifier fix that Warren made -
Oracle has a 30 character limit for identifiers and we had to shrink the
ones we had for SQL.