Clinical Import Workarounds
Still working on getting the files to import. The main problem seems to
be the fact that the whole system relies on the $login object being
valid and the $_POST and $_GET globals completely set, which is not true
when we are using the system as a service.

This update puts an IS_BATCH variable into the POST so that we can use
it as a superglobal and detect at any given point in the script whether
we are running a software service or not, and act accordlingly. There
are still fixes to be made that involve knowing this fact.