CCD Import Progress + New Test Procedure
-The CCD import process does not work because of several MySQL queries
that have to be ported to Oracle. The most notable right now is the fact
that the Oracle database (And Clicktate too!) do not allow nulls in the
hdata1 table's 'T' column, yet for some reason in the code, a query is
trying to set T to null. After investigation, it appears that this
should be correct behavior on the software's part.

I made a new test procedure (there will be MANY more to come to help us
deal with legacy code's "characteristics") that tests CCD uploads. Once
it's working we can simply port it to the inload process and
theoretically it will work without issues.