Working CCD Upload Prototype
We can now import basic users into our Oracle database!! In our test we
imported Ghandi which was successful, however when we try to import
patients with medications, there are still a few queries that fail.

Added several tests to confirm this. The tests/ccd_file_upload is what
handles all of the magic - it looks at
C:/www/clicktate/cert/sec/analytics/uploads (this is the path we will
use in production so we are emulating it) and finds a file in that
folder and will upload it and put its entry into our Oracle database.