Working Process!!
Running batch/ccd-dropbox-import/process.php will import all of the .xml
files in C:\www\clicktate\cert\sec\analytics\uploads! The BAD NEWS is
that we had to do a hacky fix in data/rec/sql/_ClientRec.php - we had to
comment out $this->throwIfDupeUid for this to work. The problem is that
when we go to upload something in our process, we look at the uploads
table, and the uploads table has a column named groupID / user group ID
or something like that, and it's always either 706 or 705. When we grab
the row from uploads and call the API for a CCD_UPLOAD call and give it
the group ID 706 or 705, it complains that a row in the CLIENTS table
already exists with the group ID 706 or 705 (the UID column in the
clients table) and therefore refuses the upload. We need to look into
this and fix it properly because duplicate user IDs in the CLIENTS table
may raise problems in the future.

We made an oracle reserved words fix in _Rec that makes sure that the
column names we are using in queries is valid for an Oracle query. In
SQL words like UID are okay to have for a column name, but not so okay
in Oracle. The fix in reservedOracleWords is to capitalize the column
name string that we are given, then if it's a reserved word add a _ at
the end. There's a hacky fix in this function that strips off the _ if
the word is UID because in the CLIENTS table there's a column named UID.
We should fix this later.