Process cleanup + Group ID / UID bug Fix
Removed some of the debugging code so the process does not echo output
and debugging code goes into a blog.txt or /logs/log.txt depending on
where the error occurred (anything that goes on with the batch process
itself goes into blog.txt in the api folder, the import process itself
that touches the system goes to logs/log.txt).

Fixed a bug where the UID and user group ID were not put into the right
columns in the CLIENTS table. They still have the duplication issue as
mentioned in the previous release because the UPLOADS table itself has
the duplicates that are forcefully-fed to the CLIENTS table.