<? to <?php fix for Analytics, Manual login being fixed
getSecurePrefix: Made an update so that 127.0.0.1 is seen as a local
host.

-Removed <? and replaced it with <?php in sec/admin*.php and many other
files in the sec/ folder.

The main thing we are working on in this release right now is getting
manual (web) logins to work. Automatic (service) logins work just fine
but since we are now using Oracle, there are several queries for the
login process that need to be ported so they work in an Oracle
environment. I suspect that virtually ALL queries will have to undergo
some type of transformation. This will be a slow process. So far manual
logins do not work but are currently the main focus.
/php/data/rec/sqlUserLogins.php, php/data/rec/sql/_SqlRec.php,
tests/ccd_file_upload and are the main files that are being worked with
for this update.

tests/loginSession_login.php is the main test that we are working on
right now, if we can get that test to pass, manual logins will work.