Client ID + Singular Database Connection
I added support for supplying client IDs in the batch process. This
allows for correct SQL queries to be generated.

Dao.php now supports single connections! Before we had no choice but to
make a new database connection for each SQL query we ran. This caused a
problem where we couldn't use Oracle's package variables because they
are reset after each database connection. This change should also help
with script speed. The comments do a good job of explaining how to do it
- add a oci_connect or mysql_connect return result to $GLOBALS['dbConn']
and make sure to unset it after all the queries in your script are done.
TODO: Somehow get this incorporated into the Dao object so that we
aren't using $GLOBALS for this.