New Tests + Improved CCD Importer
-Added a new test - ccd_file_upload_createsBLOBFile - there are
basically two operations to inloading a patient - read and create the
BLOB file from the BLOB column in the UPLOADS table, and actually
looking through that file and importing it. This test covers the former,
while ccd_file_upload_uploadsFiles.php covers the latter.

-The MsgInbox_countUnread_isNotEmpty.php test is a testing procedure for
when we get to the patients facesheet - in my local copy the query fails
because of Oracle and the page does not display correctly. Once this
query and thus this test passes, the facesheet will hopefully load.

-Fixed a bug in the Dao where incorrect values were returned for the
insert() function.

-Removed comments from several files so that it doesn't give feedback to
process.php - if process.php detects any echo'ed output, it will assume
the upload failed.