RETURNING clause support for SqlRec.php
All tests ran successfully before the commit!

-To get Dao::insert working correctly, it must return the inserted row
ID of the record that it inserts. In Oracle we have to do this by
specifying a RETURNING clause after the WHERE clause. Dao::insert does
seem to retur the correct insert ID. The problem now is that diagnoses
are not inserted into the diagnosis table - there's some kind of
permission issue going on in the code, likely a HIPPA requirement of
some sort. The good news is that patients that don't have diagnoses do
import properly.

-Removed a few echo statements that were echoing out a huge amount of
data and preventing the import process from updating the
UPLOAD.PROCESS_MSG column when an upload didn't go through and failed.
The debug code now goes to the blog / error log.