Test File Movement + CCD Import error handling + _util.php quote update
ClinicalImport_Ccd.php: Fixed an issue where the script mysteriously
dies without any errors when one of the saveAll() function calls fails.
With the try catches, the process will continue when something goes
wrong. The most notable error we get is an access not allowed (ugid,
ugi) exception (PHP application-level issue).

The same fix was done to ClinicalImport_Sql.php for the dupe check
procedures.

dao/_util.php: Updated the quote() function so that it can escape Oracle
queries with '' instead of the normal SQL \'.

Moved the tests from the unit folder to the main test folder. As Mr.
Feathers once said, tests that touch the database are not unit tests.