Oracle Query Fixes + New Patient Display Test
Auditing.php and JsonDao.php: The facesheet does not display correctly
(still needs fixes) and several relevant (hard-coded) queries were
modified when in an Oracle environment so that they run correctly. There
are likely many more queries that have to be modified so they are
compatible with Oracle - this update fixes the facesheet ones and likely
many others (courtesy of _util.php)

The getStackTrace() function in _util has a problem - when
debug_backtrace or debug_print_backtrace reads too much data, it throws
a fatal non-catchable PHP error. This rears itself when we try to load
the patients page, which tries to load 10,000+ patients. Our solution
will be to find a way to reduce the stack trace text as well as limiting
the number of results that the patient list will return (probably 35 or
100).

The new test will ensure that Javascript gets a valid object from the
server when the patients list page is loaded, but a deeper test to
ensure it's valid JSON may be needed.