HData fix + Patient page JS fixes
-PatientList_Sql: The patients page would return empty results for a
last name search because it was looking at the HDATA tables for
criteria, but since we do not encrypt data at that level, it always
returned an empty result. We now use unencrypted criteria in the fields
to find what we need.

Dao.php: The left-over echo broke Javascript.

Calendar.php, Calendar.js, pop.js: We revisit the ancient problem of
using the right event object - we have to define the event handler in
Calendar.php, and then in Calendar.js we have to grab ev.target (if
Firefox) or ev.srcElement (all other browsers) so that we have a valid
event handle to work with.

The same story goes for Pop.js - "fn" had to be defined and checked.