SqlRec Oracle Limit + PStub_Search Test + Patient List Limit
All tests except for Msg_Inbox pass this commit. Logins and the patient
page search function works.

PatientList.php: Defined a limit for the amount of rows returned.

LookupDao.php: Many queries here are hardcoded, this update is a step in
making the queries work with Oracle, but we may also want to look into
how we can not hard-code these queries. The queries in question here are
used for patient page search results.

_SqlRec.php: I thought I already commited this? The new
getOracleLimitWrapper is our answer to substituting SQL's LIMIT clause,
since Oracle does not have LIMIT. We also discovered a seam that can get
us the select fields of ANY query, which we may want to make a test for.
This can be found in the first few lines of getOracleLimitWrapper()
where we use $ci to get the SQL SELECT columns.