Oracle Query Fixes + New Test
LookupDao.php

Dao.php: Sometimes we already have a result handle from Dao::query and
we want to loop through the rows manually to do some particular
operations (for example, in LookupDao.php). The new method
fetchRowFromResource allows us to do this. The best part is that it's
completely independent of whether we're using MySQL or Oracle - this
gives us an elegant solution to hard-coded mysql_fetch_array calls that
are scattered amongst the codebase.

We had several issues today getting the facesheet to properly load the
Javascript constants. The new test here ensures that we get the correct
data from the server.

I also removed some of the debug code that could cause slow page
responses.