New Tests + Diagnoses and Medications on Facesheet
_util.php, UserDao.php: Fixed an issue where medications, diagnoses and
several other pieces of information would not load from the database
because of a mysql_fetch_array call on an Oracle resource - we used our
new Dao method from the last commit to fix this.

FacesheetDao.php: This is the main file for patient facesheet control.
The process of fetching the necessary information is under test now.
Vitals doesn't seem to even be coded in the CCD import process, even
though we have a vitals procedure here to fetch them - this is being
investigated. Allergies, medications and diagnoses load correctly now,
but doc history still needs to be fetched.

_SqlRec.php: We made a test that allows us to fetch the SELECT column
names of any given query. This required exposing the buildSqlSelectInfos
method to allow us to get the result in our hands, versus having it just
work silently in _SqlRec.php.