Vitals Import
We are working on a new addition that allows vitals from a CCD to be
imported to the vitals table in the Oracle database. We can read the
data from the CCD file and just now we parsed the XML files for the
necessary vitals information (blood pressure, weight, body temperature,
etc.). The only thing stopping us now is a user group ID issue when
attemping to parse the "insert into data_vitals" query. We have a test
made that does this same import procedure for allergies and it seems to
work fine, we just have to duplicate it for vitals.

The vitals import is a bit different than importing medications,
diagnoses and allergies - it only has ONE entry in a CCD so we do not
need to loop through entries like we do with medications and allergies
and diagnoses. Because of this, the all and from methods of
Vitals_Ci_Ccd (located in
c/patient-import/clinical-xml/ClinicalImport_Ccd.php) do not do a
foreach loop (like meds and diagnoses do) and instead works directly
with the Ccd_Section_Vitals_Sql object that it receives.

We named our base object Ccd_Section_Vitals_Sql instead of
Ccd_Section_Vitals because there is already a procedure in the codebase
that does some kind of black magic with the vitals (probably for
displaying it on the facesheet  / working with it somewhere else
somehow) and so the name is already taken.

php/data/rec/sql/dao/Dao.php and data/rec/sql/_SqlRec.php: Error
handling which writes error info to the logs - in the past it would just
silently fail.