No more silent query failures!
Made a slight fix to the Dao so that if a query fails for whatever
reason, it will snitch it out and write some details to the error log
along with the query that failed. This will tell us whenever a query
fails to run and we can fix it accordingly. This already happens with
most of the other Dao elements (selects, inserts, etc.) - the ones added
here weren't already in place yet.